Limit CCFF sections to 4GiB offsets/sizes
Previously, CCFF files used a `u64` to represent the offset/size of a
section. However, 32-bit systems cannot store that amount of data (at
least, not easily).

After considering this, I decided to limit the size to a `u32` (about
4GiB) as it is unlikely a user would need to store more than 4GiB of
data in a single section or container file, and otherwise they can use
segmentation of some sort.

This commit also resolves a few clippy warnings that were not caught as
calypso_filety was not in the module tree.