calypso_error: anyhow->eyre, non-exh. enum, 2.0.0
I've found in recent uses of eyre over anyhow in my projects that it
tends to be a bit nicer. Due to this I've switched from eyre to anyhow
within `calypso_error`.

I've also made CalError #[non_exhaustive] so that it can be easily
updated and tweaked without having to make a major bump to
`calypso_error`.

This does mean that `calypso_error` has been bumped to 2.0.0, hence all
the collateral damage within `Cargo.toml`s.